Refactor book data management for better efficiency
As part of our efforts to improve our data management, this commit performs multiple modifications in our book data processes. It removes the API/personalize.js file to refactor personalization functions, deletes some old data from the books and their collection in the database, updates the book's table structure to include a unique Google Books ID as well as description text, and transforms the publishedDate to a timestamp type for better data handling. In terms of the backend, it removes the addBookToDb endpoint since book data will be now managed differently, along with its related import in the viewBookPage Vue component, and adjusts how a new book is added in the BookDAOService, to match the modified book's table structure. Finally, the BookDAO entity changes the type of the publishedDate field to Date, and includes the fields googleBooksId and description.

By streamlining our data processes, these changes aim to make the book recommendation engine more efficient and robust.