Refactor recommendation logic and Optimise Vuex state management
Moved recommendation logic from `ViewBookPage.vue` component to Vuex Store. The recommendation data is now retrieved by a Vuex action `updateRecommendations` and stored in the Vuex state which then can be accessed using a getter `getRecommendations`. Earlier, the `getRecommendations` method in the Vue component was responsible for fetching and storing data which was violating the robust state management offered by Vuex. Separation of concern and consistency for data handling is now ensured. `fetchRecommendations` is the method which triggers Vuex action to update data. Also, ensured consistency in naming conventions for properties in book data.