Commit message: Implemented environment variables in Docker
Docker environment variables have been set up for the VUE_APP_RECOMMENDATIONS_FEATURE. The changes include a switch in usage from the process.env to window._env_ for the VUE_APP_RECOMMENDATIONS_FEATURE in the ViewBookPage.vue file. This change was necessary for fetching recommendations. Additionally, a start-up script was created and added that replaces placeholders with environment variables. Nginx starts after the execution of this script. This change allows docker to dynamically adapt the value of the VUE_APP_RECOMMENDATIONS_FEATURE at runtime.